+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 ### Fixed
 - Fix New Map settings being preserved when switching projects.
 - Fix scripting API callback `onMapResized` not triggering.
+- Fix crash when importing AdvanceMap metatiles while `enable_triple_layer_metatiles` is enabled.
 
 ## [5.0.0] - 2022-10-30
 ### Breaking Changes